style: rustfmt/clippy sweep across frankenlibc crates
Mechanical lint hygiene across ~50 files in frankenlibc-abi, -bench, -core
and -harness; no behavior change:

- rustfmt re-wraps: long call chains, multi-line tuple returns, assert!
  messages (math/exp.rs, math/float32.rs), closure bodies (fnmatch's
  casefold eq, sort's PCG step), bench_with_input closures in string_bench
  and iconv_bench, import lists.
- clippy idiom fixes: let-chains replacing nested if-let+condition
  (stdio_abi limit check), `wmemchr(..)?;` replacing
  if-is_none-return-None (string/wide.rs), redundant cast drop in
  sort.rs's corpus builder, unused-arg renames.
- string/regex.rs: fold the VM's notbol/noteol flags into a VmAnchors
  struct and the visited-bitmap + generation pair into ClosureState,
  threading them through the epsilon-closure calls instead of long
  positional argument lists (clippy too_many_arguments); literal prefix
  search now uses `find_literal(..)?` directly. Same algorithm, same
  outputs.
- differential-probe and harness test files: formatting-only re-wraps of
  probe tables, assert messages and fixture literals.

+++ crates/frankenlibc-core/src/stdlib/sort.rs
@@ -978,7 +978,9 @@ mod sort_variant_tests {
 
     /// Deterministic LCG so the corpus is fixed (no `rand`, no clock).
     fn lcg(state: &mut u64) -> u64 {
-        *state = state.wrapping_mul(6364136223846793005).wrapping_add(1442695040888963407);
+        *state = state
+            .wrapping_mul(6364136223846793005)
+            .wrapping_add(1442695040888963407);
         *state
     }
 
@@ -997,7 +999,7 @@ mod sort_variant_tests {
             // low cardinality (mod 16)
             corpus.push((0..n as u32).map(|i| (i * 7) % 16).collect());
             // sawtooth
-            corpus.push((0..n as u32).map(|i| (i % 50) as u32).collect());
+            corpus.push((0..n as u32).map(|i| i % 50).collect());
             // organ pipe: 0..n/2 then n/2..0
             corpus.push(
                 (0..n)
